GetWeather/requirements.txt
requests
python-telegram-bot
beautifulSoup4